+++ source/git-promise.js
@@ -53,7 +53,7 @@ const gitExecutorProm = (args, retryCount) => {
           winston.info(`git executing: ${args.repoPath} ${args.commands.join(' ')}`);
         let rejectedError = null;
         let isCompleted = false;
-        let timer = null;
+        let timeoutTimer = null;
         let stdout = '';
         let stderr = '';
         let env = JSON.parse(JSON.stringify(process.env));
@@ -66,10 +66,10 @@ const gitExecutorProm = (args, retryCount) => {
         };
         const gitProcess = child_process.spawn(gitBin, args.commands, procOpts);
         if (args.timeout) {
-          timer = setTimeout(() => {
+          timeoutTimer = setTimeout(() => {
             if (isCompleted) return;
             isCompleted = true;
-            timer = null;
+            timeoutTimer = null;
             winston.warn(`command timedout: ${args.commands.join(' ')}\n`);
             gitSem.signal();
             gitProcess.kill('SIGINT');
@@ -92,9 +92,9 @@ const gitExecutorProm = (args, retryCount) => {
         gitProcess.on('close', (code) => {
           if (isCompleted) return;
           isCompleted = true;
-          if (timer) {
-            clearTimeout(timer);
-            timer = null;
+          if (timeoutTimer) {
+            clearTimeout(timeoutTimer);
+            timeoutTimer = null;
           }
           if (config.logGitCommands)
             winston.info(